# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDR Center Uzbekistan API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface Service {
  id: number;
  nameUz: string;
  nameEn: string;
  nameRu: string;
  descriptionUz: string;
  descriptionEn: string;
  descriptionRu: string;
  price?: number | null;
  imageUrl?: string | null;
  category?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface ServiceInput {
  nameUz: string;
  nameEn: string;
  nameRu: string;
  descriptionUz: string;
  descriptionEn: string;
  descriptionRu: string;
  price?: number | null;
  imageUrl?: string | null;
  category?: string | null;
  sortOrder?: number;
}

export interface Product {
  id: number;
  nameUz: string;
  nameEn: string;
  nameRu: string;
  descriptionUz: string;
  descriptionEn: string;
  descriptionRu: string;
  price: number;
  imageUrl?: string | null;
  category: string;
  inStock: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface ProductInput {
  nameUz: string;
  nameEn: string;
  nameRu: string;
  descriptionUz: string;
  descriptionEn: string;
  descriptionRu: string;
  price: number;
  imageUrl?: string | null;
  category: string;
  inStock?: boolean;
  sortOrder?: number;
}

export interface Course {
  id: number;
  nameUz: string;
  nameEn: string;
  nameRu: string;
  descriptionUz: string;
  descriptionEn: string;
  descriptionRu: string;
  price: number;
  durationDays: number;
  imageUrl?: string | null;
  level: string;
  sortOrder: number;
  createdAt: string;
}

export interface CourseInput {
  nameUz: string;
  nameEn: string;
  nameRu: string;
  descriptionUz: string;
  descriptionEn: string;
  descriptionRu: string;
  price: number;
  durationDays: number;
  imageUrl?: string | null;
  level?: string;
  sortOrder?: number;
}

export interface GalleryItem {
  id: number;
  titleUz: string;
  titleEn: string;
  titleRu: string;
  beforeImage: string;
  afterImage: string;
  category: string;
  carBrand?: string | null;
  createdAt: string;
}

export interface GalleryItemInput {
  titleUz: string;
  titleEn: string;
  titleRu: string;
  beforeImage: string;
  afterImage: string;
  category?: string;
  carBrand?: string | null;
}

export interface Review {
  id: number;
  author: string;
  rating: number;
  textUz: string;
  textEn: string;
  textRu: string;
  carBrand?: string | null;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  author: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  textUz: string;
  textEn: string;
  textRu: string;
  carBrand?: string | null;
  avatarUrl?: string | null;
}

export interface Booking {
  id: number;
  name: string;
  phone: string;
  email?: string | null;
  serviceId?: number | null;
  message?: string | null;
  status: string;
  createdAt: string;
}

export interface BookingInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 9 */
  phone: string;
  email?: string | null;
  serviceId?: number | null;
  message?: string | null;
}

export type BookingStatusUpdateStatus =
  (typeof BookingStatusUpdateStatus)[keyof typeof BookingStatusUpdateStatus];

export const BookingStatusUpdateStatus = {
  new: "new",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}

export type GetProductsParams = {
  category?: string;
};

export type GetGalleryParams = {
  category?: string;
};

export interface Article {
  id: number;
  titleUz: string;
  titleEn: string;
  titleRu: string;
  contentUz: string;
  contentEn: string;
  contentRu: string;
  imageUrl?: string | null;
  authorId?: string | null;
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ArticleInput {
  titleUz: string;
  titleEn: string;
  titleRu: string;
  contentUz: string;
  contentEn: string;
  contentRu: string;
  imageUrl?: string | null;
  authorId?: string | null;
  publishedAt?: string | null;
}