# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDR Center Uzbekistan API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  name: string;
  phone: string;
  email?: string | null;
  serviceId?: number | null;
  message?: string | null;
  status: string;
  createdAt: Date;
}