# lib/api-zod/src/generated/types/bookingStatusUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDR Center Uzbekistan API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatusUpdateStatus } from "./bookingStatusUpdateStatus";

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}