# lib/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PDR Center Uzbekistan API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: Get all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a single course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: Get all reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Add a review (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a review by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Not found
    put:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Review updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Not found
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete review (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get all bookings (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Submit a booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
  /bookings/{id}/status:
    put:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Booking status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /articles:
    get:
      operationId: getArticles
      tags: [articles]
      summary: Get all articles
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create an article (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get a single article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    put:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Course:
      type: object
      properties:
        id:
          type: integer
        nameUz:
          type: string
        nameEn:
          type: string
        nameRu:
          type: string
        descriptionUz:
          type: string
        descriptionEn:
          type: string
        descriptionRu:
          type: string
        price:
          type: integer
        durationDays:
          type: integer
        imageUrl:
          type: string
          nullable: true
        level:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameUz
        - nameEn
        - nameRu
        - descriptionUz
        - descriptionEn
        - descriptionRu
        - price
        - durationDays
        - level
        - sortOrder
        - createdAt
    CourseInput:
      type: object
      properties:
        nameUz:
          type: string
        nameEn:
          type: string
        nameRu:
          type: string
        descriptionUz:
          type: string
        descriptionEn:
          type: string
        descriptionRu:
          type: string
        price:
          type: integer
        durationDays:
          type: integer
        imageUrl:
          type: string
          nullable: true
        level:
          type: string
        sortOrder:
          type: integer
      required:
        - nameUz
        - nameEn
        - nameRu
        - descriptionUz
        - descriptionEn
        - descriptionRu
        - price
        - durationDays
...
    Review:
      type: object
      properties:
        id:
          type: integer
        author:
          type: string
        rating:
          type: integer
        textUz:
          type: string
        textEn:
          type: string
        textRu:
          type: string
        carBrand:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - author
        - rating
        - textUz
        - textEn
        - textRu
        - createdAt
    ReviewInput:
      type: object
      properties:
        author:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        textUz:
          type: string
        textEn:
          type: string
        textRu:
          type: string
        carBrand:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required:
        - author
        - rating
        - textUz
        - textEn
        - textRu
    Booking:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        serviceId:
          type: integer
          nullable: true
        message:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - status
        - createdAt
    BookingInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 9
        email:
          type: string
          nullable: true
        serviceId:
          type: integer
          nullable: true
        message:
          type: string
          nullable: true
      required:
        - name
        - phone
    BookingStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, in_progress, completed, cancelled]
      required:
        - status
    Article:
      type: object
      properties:
        id:
          type: integer
        titleUz:
          type: string
        titleEn:
          type: string
        titleRu:
          type: string
        contentUz:
          type: string
        contentEn:
          type: string
        contentRu:
          type: string
        imageUrl:
          type: string
          nullable: true
        authorId:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - titleUz
        - titleEn
        - titleRu
        - contentUz
        - contentEn
        - contentRu
        - createdAt
        - updatedAt
    ArticleInput:
      type: object
      properties:
        titleUz:
          type: string
        titleEn:
          type: string
        titleRu:
          type: string
        contentUz:
          type: string
        contentEn:
          type: string
        contentRu:
          type: string
        imageUrl:
          type: string
          nullable: true
      required:
        - titleUz
        - titleEn
        - titleRu
        - contentUz
        - contentEn
        - contentRu